allow usage without user having to set PYTHONPATH
Using the code generator currently requires the user to manipulate the
PYTHONPATH environment variable to find both the dots module and the
configuration file. Since it is not always possible to easily set the
PYTHONPATH (e.g. in cross-platform scenarios), a user should also be
able to use the generator without it.

Since the dots module is always located in a location relative to the
generator, its path can be added automatically when the script is
loaded. For specifying the config file, this commit also adds the
option of using a path instead of a name.